feat: exteded main component to department pages
